Fixed getMessageList bugs

+++ project/radov/common/Helper.h
@@ -12,7 +12,7 @@ class Helper
 public:
     static BuffersVector addSize(ByteBufferPtr buffer);
     static EndBuffer makeEndBuffer(BuffersVector buffers);
-    static uint16_t getSize(uint16_t left, uint16_t right);
+    static uint16_t getSize(uint8_t left, uint8_t right);
     static bool parseFromPostgres(const pqxx::tuple& data, User& user);
     static bool parseChatMessages(const pqxx::tuple& data, ChatMessage& chatMessage);
     static void prependCommand(Commands command, std::string& message);
